+++ kytos/utils/napps.py
@@ -37,11 +37,23 @@ class NAppsManager:
         """
         self._controller = controller
         self._config = KytosConfig().config
-        self._installed = Path(self._config.get('napps', 'installed_path'))
-        self._enabled = Path(self._config.get('napps', 'enabled_path'))
+        self._kytos_api = self._config.get('kytos', 'api')
+        self._load_kytos_configuration()
 
         self.user = None
         self.napp = None
+
+    def _load_kytos_configuration(self):
+        """Request current configurations loaded by Kytos instance."""
+        uri = self._kytos_api + 'kytos/config/'
+        try:
+            options = json.loads(urllib.request.urlopen(uri).read())
+        except urllib.error.URLError:
+            print('Kytos is not running.')
+            sys.exit()
+
+        self._installed = Path(options.get('installed_napps'))
+        self._enabled = Path(options.get('napps'))
 
     def set_napp(self, user, napp):
         """Set info about NApp.
@@ -126,13 +138,21 @@ class NAppsManager:
 
     def disable(self):
         """Disable a NApp if it is enabled."""
-        enabled = self._enabled / self.user / self.napp
+        enabled = self.enabled_dir()
         try:
             enabled.unlink()
             if self._controller is not None:
                 self._controller.unload_napp(self.user, self.napp)
         except FileNotFoundError:
             pass  # OK, it was already disabled
+
+    def enabled_dir(self):
+        """Return the enabled dir from current napp."""
+        return self._enabled / self.user / self.napp
+
+    def installed_dir(self):
+        """Return the installed dir from current napp."""
+        return self._installed / self.user / self.napp
 
     def enable(self):
         """Enable a NApp if not already enabled.
@@ -141,8 +161,8 @@ class NAppsManager:
             FileNotFoundError: If NApp is not installed.
             PermissionError: No filesystem permission to enable NApp.
         """
-        enabled = self._enabled / self.user / self.napp
-        installed = self._installed / self.user / self.napp
+        enabled = self.enabled_dir()
+        installed = self.installed_dir()
 
         if not installed.is_dir():
             raise FileNotFoundError('Install NApp {} first.'.format(
@@ -167,7 +187,7 @@ class NAppsManager:
     def uninstall(self):
         """Delete code inside NApp directory, if existent."""
         if self.is_installed():
-            installed = self._installed / self.user / self.napp
+            installed = self.installed_dir()
             if installed.is_symlink():
                 installed.unlink()
             else:
@@ -217,7 +237,7 @@ class NAppsManager:
             FileNotFoundError: If NApp is not found.
         """
         folder = self._get_local_folder()
-        installed = self._installed / self.user / self.napp
+        installed = self.installed_dir()
         self._check_module(installed.parent)
         installed.symlink_to(folder.resolve())